feat: larger input trigger click area
